refactor: use correct filename for AppearanceMenu
